+++ DTLiving/Bridge/ImageProcessing/VideoSobelEdgeDetectionFilter.h
@@ -12,7 +12,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface VideoSobelEdgeDetectionFilter : VideoFilter
 
-@property(nonatomic, assign) CGFloat edgeStrength;
+@property(nonatomic, assign) float edgeStrength;
 
 - (instancetype)init;
